# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pixel-BD is an open-source online platform where anyone can upload, share, and showcase their pixel art creations with the community.",
        "title": "PixelDB",
        "contact": {
            "name": "Lukas Haible",
            "email": "lukas.haible@web.de"
        },
        "version": "0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "This is the base route. You can check, if the API is online.",
                "tags": [
                    "Base"
                ],
                "summary": "Hello",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login with your credentials, to get your User Data and your JWT-Token",
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AuthResponse"
                        }
                    }
                }
            }
        },
        "/pixelart": {
            "get": {
                "description": "Returns a paginated list of pixel art",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PixelArt"
                ],
                "summary": "Get pixel art list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Size of each Page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of pixel art",
                "tags": [
                    "PixelArt"
                ],
                "summary": "Upload PixelArt",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PixelArt-File",
                        "name": "pixelart",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UploadFileResponse"
                        }
                    }
                }
            }
        },
        "/pixelart/{pixelArtID}": {
            "get": {
                "description": "Returns the infos for a specific pixel art by ID",
                "tags": [
                    "PixelArt"
                ],
                "summary": "Get PixelArt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PixelArt ID",
                        "name": "pixelArtID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pixelart/{pixelArtID}/picture": {
            "get": {
                "description": "Returns the image for a specific pixel art by ID",
                "produces": [
                    "image/jpeg",
                    "image/png"
                ],
                "tags": [
                    "PixelArt"
                ],
                "summary": "Get PixelArt Picture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PixelArt ID",
                        "name": "pixelArtID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the user Data, when passing your JWT-Token in the Heeader",
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new user account by accepting user details such as username, email, password, firstname and lastname.",
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "User Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.AuthResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.Data": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "Role": {
                    "type": "string"
                },
                "Username": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                },
                "Status": {
                    "type": "string",
                    "example": "Error"
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.UploadFileResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "oldFilename": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                },
                "ownerUsername": {
                    "type": "string"
                },
                "pixelArtSize": {
                    "type": "integer",
                    "format": "int64"
                },
                "pixelArtURL": {
                    "type": "string"
                }
            }
        },
        "dto.UserRequest": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "maxLength": 70,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/dto.Data"
                },
                "Message": {
                    "type": "string",
                    "example": "Created User"
                },
                "Status": {
                    "type": "string",
                    "example": "Success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}